+++ jsonmap/codegen/padGroup.h
@@ -42,11 +42,11 @@ getPadGroups(const std::vector<MotifPosition> &motifPositions, const std::vector
 std::vector<PadGroupType>
 getPadGroupTypes(const std::vector<PadGroup> &padGroups, const std::vector<PadGroupType> &padGroupTypes);
 
-std::vector<std::pair<float, float>>
-getPadSizes(const std::vector<PadGroup> &padGroups, const std::vector<std::pair<float, float>> &allPadSizes);
+std::vector<std::pair<double,double>>
+getPadSizes(const std::vector<PadGroup> &padGroups, const std::vector<std::pair<double,double>> &allPadSizes);
 
 std::vector<PadGroup> remap(const std::vector<PadGroup> &padGroups, const std::vector<PadGroupType> &padGroupTypes,
-                            const std::vector<std::pair<float, float>> &padSizes);
+                            const std::vector<std::pair<double,double>> &padSizes);
 
 std::ostream &operator<<(std::ostream &os, const PadGroup &pg);
